Enhance posture analysis API and notification system
- Introduced new endpoints: `/api/features/aggregate/add`, `/api/features/aggregate/mean`, `/api/features/aggregate/clear`, and `/api/decide/from_buffer`.
- Added support for RL and ML analysis with `/api/rl/analyze` and `/api/ml/analyze` endpoints.
- Refactored `NotificationService` to integrate with the new decision-making logic and manage notifications based on aggregated features.
- Enhanced frontend to allow users to enable notifications and select decision methods (RL or ML).
- Removed deprecated `feature_extractor.py` and integrated its functionality into `pose_service.py` for streamlined feature extraction.
