+++ app/api/routes.py
@@ -1,53 +1,30 @@
-from fastapi import APIRouter, UploadFile, File, HTTPException, WebSocket, Request
+from fastapi import APIRouter, UploadFile, File, HTTPException, WebSocket
 from fastapi.responses import JSONResponse
 from starlette.websockets import WebSocketDisconnect
 
-from app.api.schemas import AnalyzeResponse, FeatureExtractionResponse, Notification, NotificationSeverity, RLAgentState
+from app.api.schemas import FeatureExtractionResponse, Notification, NotificationSeverity, RLAgentState, ModelAnalysisRequest, ModelAnalysisResponse, FeatureVector
 from app.services.pose_service import PoseService, PoseServiceUnavailable
-from app.services.feature_extractor import PostureFeatureExtractor
 from app.services.notification_service import NotificationService
+from app.services.feature_buffer import FeatureBuffer
+from app.services.rl_service import RLService
 
 router = APIRouter()
 
 _notification_ws_clients = set()
 _last_notification: Notification | None = None
+_buffer = FeatureBuffer()
 
 
 @router.get("/health", tags=["system"])
 async def health():
     return {"status": "ok"}
 
-
-@router.post("/analyze", response_model=AnalyzeResponse, tags=["analysis"])
-async def analyze(image: UploadFile = File(...)):
-    try:
-        image_bytes = await image.read()
-        result = PoseService.get_instance().analyze_image(image_bytes)
-        return AnalyzeResponse.model_validate(result)
-    except PoseServiceUnavailable as e:
-        raise HTTPException(status_code=503, detail=str(e))
-    except Exception as e:  # noqa: BLE001 - surface error to client for now
-        raise HTTPException(status_code=400, detail=f"Failed to analyze image: {e}")
-
-
 @router.post("/features", response_model=FeatureExtractionResponse, tags=["analysis"])
 async def extract_features(image: UploadFile = File(...)):
     try:
         image_bytes = await image.read()
-        extractor = PostureFeatureExtractor()
-        result = extractor.extract_features(image_bytes)
-        # Non-blocking notification check (fire-and-forget); do not raise if it fails
-        try:
-            features = result.get("features")
-            should_notify = NotificationService.get_instance().maybe_notify(features)
-            print(
-                "[routes:/features] features present=",
-                bool(features),
-                "should_notify=",
-                should_notify,
-            )
-        except Exception as e:
-            print(f"[routes:/features] notification check failed: {e}")
+        # Use PoseService for both pose and features
+        result = PoseService.get_instance().extract_features(image_bytes)
         return FeatureExtractionResponse.model_validate(result)
     except PoseServiceUnavailable as e:
         raise HTTPException(status_code=503, detail=str(e))
@@ -55,48 +32,6 @@ async def extract_features(image: UploadFile = File(...)):
         raise HTTPException(status_code=400, detail=f"Failed to extract features: {e}")
 
 
-@router.websocket("/ws/analyze")
-async def ws_analyze(websocket: WebSocket):
-    await websocket.accept()
-    try:
-        service = PoseService.get_instance()
-    except PoseServiceUnavailable as e:
-        try:
-            await websocket.send_json({"error": str(e)})
-        finally:
-            # Closing once with a normal code; avoid double close
-            await websocket.close(code=1013)
-        return
-
-    await websocket.send_json({"status": "ready"})
-
-    try:
-        while True:
-            try:
-                message = await websocket.receive()
-                if message["type"] == "websocket.disconnect":
-                    break
-                if "text" in message and message["text"] is not None:
-                    text = message["text"].strip().lower()
-                    if text == "ping":
-                        await websocket.send_json({"pong": True})
-                    continue
-                frame = message.get("bytes")
-                if not frame:
-                    continue
-                try:
-                    result = service.analyze_image(frame)
-                    await websocket.send_json(result)
-                except Exception as e:  # noqa: BLE001
-                    await websocket.send_json({"error": f"analyze_failed: {e}"})
-            except WebSocketDisconnect:
-                break
-    except Exception:
-        # If something unexpected happens, attempt a graceful close
-        try:
-            await websocket.close()
-        except Exception:
-            pass
 
 
 @router.get("/reminder", response_model=Notification, tags=["notifications"]) 
@@ -183,5 +118,66 @@ async def notifications_last():
 
 @router.get("/rl/state", response_model=RLAgentState, tags=["rl"])
 async def rl_state():
-    svc = NotificationService.get_instance()
-    return RLAgentState.model_validate(svc.get_rl_state())
+    # delegate to RL service
+    return RLAgentState.model_validate(RLService.get_instance().get_state())
+
+
+# RL analyze route (same contract as future ML analyze route)
+@router.post("/rl/analyze", response_model=ModelAnalysisResponse, tags=["rl"])
+async def rl_analyze(payload: ModelAnalysisRequest):
+    resp = RLService.get_instance().analyze(payload)
+    return resp
+
+
+# Future ML analyze stub with the same contract
+@router.post("/ml/analyze", response_model=ModelAnalysisResponse, tags=["ml"])
+async def ml_analyze_stub(payload: ModelAnalysisRequest):
+    return ModelAnalysisResponse(should_notify=False, reason="ml-stub")
+
+
+# Decide using the server buffer (mean over window), then notify
+@router.post("/decide/from_buffer", response_model=ModelAnalysisResponse, tags=["decision"])
+async def decide_from_buffer(method: str = "rl"):
+    try:
+        mean_features = _buffer.mean()
+    except Exception:
+        mean_features = None
+    if not mean_features:
+        return ModelAnalysisResponse(should_notify=False, reason="no-features")
+    # Toggle RL training according to method
+    try:
+        from app.services.rl_service import RLService as _RS
+        _RS.get_instance().set_training_enabled(method == "rl")
+    except Exception:
+        pass
+    NotificationService.get_instance().decide_and_notify(mean_features, method=method)
+    # Also return model decision for clients who call this endpoint
+    if method == "rl":
+        resp = RLService.get_instance().analyze(ModelAnalysisRequest(features=FeatureVector(**mean_features)))
+    else:
+        resp = ModelAnalysisResponse(should_notify=False, reason="ml-stub")
+    return resp
+
+# Feature aggregation management endpoints
+@router.post("/features/aggregate/add", tags=["aggregation"])
+async def features_aggregate_add(payload: FeatureVector):
+    try:
+        _buffer.add(payload.model_dump())
+    except Exception:
+        pass
+    return {"ok": True}
+
+
+@router.get("/features/aggregate/mean", tags=["aggregation"])
+async def features_aggregate_mean():
+    mean = _buffer.mean()
+    return {"features": mean}
+
+
+@router.post("/features/aggregate/clear", tags=["aggregation"])
+async def features_aggregate_clear():
+    from app.services.feature_buffer import FeatureBuffer as _FB
+    # reinitialize buffer to clear it
+    global _buffer
+    _buffer = _FB(window_seconds=_buffer.window_seconds)
+    return {"ok": True}
